Added Files
Live-Video-Testing: The primary code comprises two segments. The first segment is activated to optimize project performance, while the second, which is commented, is intended exclusively for testing purposes. The latter provides a display of both the live feed and the image masking view.

Training.py: This script is employed for model training. It involves specifying the path to an image folder, which should be modified if there is a need to train a new model. The existing model has achieved an accuracy level of 0.96.

traffic_sign_model.h5: This file houses the trained model, boasting an impressive accuracy rate of 0.96.